[Refactor/#102] 지도 API 수정

+++ app/src/main/java/com/example/bisit/data/model/payment/PaymentConfirmRequest.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class PaymentConfirmRequest(
     @SerializedName("paymentKey") val paymentKey: String,
     @SerializedName("orderId") val orderId: String,
-    @SerializedName("amount") val amount: Int
+    @SerializedName("amount") val amount: Long
 )
